Калькулятор Math.toIntExact

+++ src/main/java/faang/school/godbless/BJS2_796/Calculator.java
@@ -2,5 +2,5 @@ package faang.school.godbless.BJS2_796;
 
 @FunctionalInterface
 public interface Calculator {
-    int operate (int a, int b);
+    long operate (int a, int b);
 }
